Kimi_Agent_Quantum Linalg Teacher & Solver/requirements.txt
openai>=0.27.0
streamlit>=1.28.0
python-dotenv>=1.0.0
python-dotenv
